some refactoring, adding finals

+++ src/main/java/PriceStrategies/PriceStrategy.java
@@ -4,5 +4,5 @@ package PriceStrategies;
  * Created by chris on 24.01.2018.
  */
 public interface PriceStrategy {
-	long getPrice(int itemCount);
+	long getPrice(long itemCount);
 }
